Bug #561996 add sprd secboot 6.0mr1_prime_16b
[root cause  ] add secboot
[changes     ] add secboot
[side effects] noside customer not use
[self test   ]<board, Test cases summary>:<sp7730sw,boot ok>
[reviewers   ] adam sun
[change_type ] bug fix
[tag_product ] common

+++ packimage/signimage/sprd/common/verify/inc/sprdsec_header.h
@@ -8,7 +8,7 @@ typedef signed int int32_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned long int uint64_t;
 
 #define MAGIC_SIZE 8
 #define VERSION_EBL 0
